Refactoring DI
* group all modules into package
* fix access modifiers
