+++ app/src/main/java/com/github/mik629/android/fundamentals/di/AuthInterceptor.kt
@@ -1,0 +1,21 @@
+package com.github.mik629.android.fundamentals.di
+
+import com.github.mik629.android.fundamentals.BuildConfig
+import okhttp3.Interceptor
+import okhttp3.Response
+
+internal class AuthInterceptor : Interceptor {
+    override fun intercept(chain: Interceptor.Chain): Response {
+        val newUrl = chain.request().url
+            .newBuilder()
+            .addQueryParameter("api_key", BuildConfig.API_KEY)
+            .build()
+
+        val newRequest = chain.request()
+            .newBuilder()
+            .url(newUrl)
+            .build()
+
+        return chain.proceed(newRequest)
+    }
+}
