+++ app/src/main/java/com/github/mik629/android/fundamentals/di/modules/AppModule.kt
@@ -1,4 +1,4 @@
-package com.github.mik629.android.fundamentals.di
+package com.github.mik629.android.fundamentals.di.modules
 
 import android.app.Application
 import android.content.Context
@@ -8,7 +8,7 @@ import timber.log.Timber
 import javax.inject.Singleton
 
 @Module
-object AppModule {
+internal object AppModule {
     @Provides
     @Singleton
     fun provideAppContext(application: Application): Context = application.applicationContext
