+++ app/src/main/java/com/github/mik629/android/fundamentals/di/modules/NetworkModule.kt
@@ -1,13 +1,12 @@
-package com.github.mik629.android.fundamentals.di
+package com.github.mik629.android.fundamentals.di.modules
 
 import com.github.mik629.android.fundamentals.BuildConfig
 import com.github.mik629.android.fundamentals.data.network.ServerApi
+import com.github.mik629.android.fundamentals.di.AuthInterceptor
 import com.squareup.moshi.Moshi
 import dagger.Module
 import dagger.Provides
-import okhttp3.Interceptor
 import okhttp3.OkHttpClient
-import okhttp3.Response
 import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.converter.moshi.MoshiConverterFactory
@@ -18,11 +17,9 @@ internal class NetworkModule {
     @Provides
     @Singleton
     fun provideOkHttpClient(): OkHttpClient {
-        val loggingInterceptor = HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BODY
-        }
+        val loggingInterceptor = HttpLoggingInterceptor()
+        loggingInterceptor.level = HttpLoggingInterceptor.Level.BODY
         return OkHttpClient().newBuilder()
-            .addInterceptor(loggingInterceptor)
             .addNetworkInterceptor(loggingInterceptor)
             .addInterceptor(AuthInterceptor())
             .build()
@@ -44,20 +41,4 @@ internal class NetworkModule {
     @Singleton
     fun provideApi(retrofit: Retrofit): ServerApi = retrofit.create(ServerApi::class.java)
 
-}
-
-private class AuthInterceptor : Interceptor {
-    override fun intercept(chain: Interceptor.Chain): Response {
-        val newUrl = chain.request().url
-            .newBuilder()
-            .addQueryParameter("api_key", BuildConfig.API_KEY)
-            .build()
-
-        val newRequest = chain.request()
-            .newBuilder()
-            .url(newUrl)
-            .build()
-
-        return chain.proceed(newRequest)
-    }
 }
